refactor(domain-context): consolidate bound store access
Constraint: Preserve descriptor-bound, fail-closed store semantics and polling limits
Rejected: Keep bound helpers split across writer and security | obscures access ownership
Confidence: high
Scope-risk: narrow
Directive: Keep descriptor-relative domain access in the bound store module
Tested: 73 targeted tests; compileall; Ruff; git diff --check
Not-tested: Full unittest discovery
